UCS/STRING: Fix definition of UCS_ULUNITS_INF

+++ src/ucs/sys/string.h
@@ -28,7 +28,7 @@ BEGIN_C_DECLS
 
 /* the numeric value of "infinity" */
 #define UCS_MEMUNITS_INF    ((size_t)-1)
-#define UCS_ULUNITS_INF     ((size_t)-1)
+#define UCS_ULUNITS_INF     ((unsigned long)-1)
 
 /* value which specifies "auto" for a numeric variable */
 #define UCS_MEMUNITS_AUTO   ((size_t)-2)
